+++ Emulator/FileSystems/FSFileHeaderBlock.h
@@ -27,7 +27,7 @@ struct FSFileHeaderBlock : FSBlock {
     FSItemType itemType(isize byte) const override;
     ErrorCode check(isize pos, u8 *expected, bool strict) const override;
     void dump() const override;
-    u32 checksumLocation() const override { return 5; }
+    isize checksumLocation() const override { return 5; }
 
     ErrorCode exportBlock(const char *path) override;
