+++ Emulator/FileSystems/FSRootBlock.h
@@ -27,7 +27,7 @@ struct FSRootBlock : FSBlock {
     FSBlockType type() const override { return FS_ROOT_BLOCK; }
     FSItemType itemType(isize byte) const override;
     ErrorCode check(isize pos, u8 *expected, bool strict) const override;
-    u32 checksumLocation() const override { return 5; }
+    isize checksumLocation() const override { return 5; }
  
     u32 getBmBlockRef(isize nr) const            { return get32(nr-49);        }
     void setBmBlockRef(isize nr, u32 ref)        {        set32(nr-49, ref);   }
